Memory management and variable arguments
Create snippets for the use of malloc/calloc/realloc, also pairing the
first two with the corresponding free. Also add a similar snippet for
managing variable number of arguments in a function, from the creation
in the va_list to the destruction with va_end.

Also add some small snippets like scanf() or seed_rand()

Small changes in some other snippets, like a new alias for
_Static_assert(), no comma in printf to be like fprintf.