c,cpp: Remove const from argv of main
The parameter argv of main with type 'char const * []' is incompatible
with libc getopt() and long variants, which expects a 'char * const []'.

Hence, we only have to getopt()-compatible variants for the type of
argv:
  1. char * []
  2. char * const []

Variant 1 is the one mentioned in the ISO C99 standard (§5.1.2.2.1) and
is the one used by the man page of getopt() or by cppreference.com on
the main function.